+++ changelog.md
@@ -1,3 +1,4 @@
 ## Fabric & NeoForge
 
+- prevent item loss when inventory mods fail during capacity checks [**#106**](https://github.com/btwonion/magnetic/issues/106)
 - fix flame kills not being processed [**#105**](https://github.com/btwonion/magnetic/issues/105)